refactor(mdx): integrate TOC generation directly into generated documents and update related types
