+++ packages/mdx/src/get-mdx-component.ts
