+++ packages/mdx/src/index.ts
@@ -1,2 +1,2 @@
+export * from './get-mdx-component'
 export * from './types'
-export * from './utils'
