--- packages/mdx/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './get-mdx-component'
-export * from './get-toc'
